+++ changelog.md
@@ -16,6 +16,7 @@
 - Fixed alternative handler file checks when using Deno; fixes #1022
 - Fixed formatting in unknown `@http` userland error state
 - Fixed middleware handling `ARC_SANDBOX_ENABLE_CORS`; thanks @neilhoff!
+- Fixes serving bare `@static` (i.e. S3-only with no `@http`) apps; fixes #1031, thanks @dam!
 
 ---